requirements.txt
# Python dependencies with known vulnerabilities
django==2.2.0
requests==2.20.0
pyyaml==3.13
jinja2==2.10
flask==1.0.0
cryptography==2.6.1
pillow==6.2.0
